fix for dublicated digital signature aspect, when a document is being copied in alfresco

+++ esign-cert-repo/src/main/java/es/keensoft/alfresco/behaviour/TransactionalDeleteBehaviour.java
@@ -114,7 +114,9 @@ public class TransactionalDeleteBehaviour implements NodeServicePolicies.BeforeD
 	                     
 	                    @Override
 	                    public Void execute() throws Throwable {
-	                        nodeService.deleteNode(nodeToBeReviewed);
+	                        if(nodeService.exists(nodeToBeReviewed)) {
+	                    		nodeService.deleteNode(nodeToBeReviewed);
+	                    	}
 	                        return null;
 	                    }
 	                };
